feat: syncing tournaments improvements

+++ apps/worker/sync/src/app/processors/sync-events-v2/tournament/processors/draw.processor.ts
@@ -1,7 +1,6 @@
 import { DrawTournament, RankingSystem, SubEventTournament } from '@badman/backend-database';
 import { Sync, SyncQueue, TransactionManager } from '@badman/backend-queue';
 import { VisualService, XmlDrawTypeID } from '@badman/backend-visual';
-import { JobsService } from '@badman/frontend-queue';
 import { DrawType } from '@badman/utils';
 import { InjectQueue, Process, Processor } from '@nestjs/bull';
 import { Logger } from '@nestjs/common';
@@ -25,7 +24,7 @@ export class DrawTournamentProcessor {
       // transaction
       transactionId: string;
 
-      // provide or direed
+      // provide or derived
       eventCode: string;
       subEventId: string;
       rankingSystemId: string;
@@ -35,11 +34,27 @@ export class DrawTournamentProcessor {
       drawCode: number;
 
       // options
-      updateMatches: boolean;
-      updateStanding: boolean;
+      options: {
+        deleteDraw?: boolean;
+        deleteMatches?: boolean;
+        deleteStandings?: boolean;
+
+        updateMatches?: boolean;
+        updateStanding?: boolean;
+      };
+
+      // from parent
+      games: { id: string; visualCode: string }[];
     }>,
   ): Promise<void> {
     const transaction = await this._transactionManager.getTransaction(job.data.transactionId);
+
+    const options = {
+      // update when we delete the draw (unless specified)
+      updateMatches: job.data.options?.deleteDraw || false,
+      updateStanding: job.data.options?.deleteDraw || false,
+      ...job.data.options,
+    };
 
     let draw: DrawTournament;
     if (job.data.drawId) {
@@ -90,11 +105,28 @@ export class DrawTournamentProcessor {
     // delete the data and reuse the guid
     const drawId = draw?.id;
     const drawCode = draw?.visualCode || job.data.drawCode.toString();
-    let existed = false;
-    if (draw) {
+    const existing = {
+      existed: false,
+      games: job.data?.games || [],
+    };
+    if (draw && options.deleteDraw) {
       this.logger.debug(`Deleting draw ${draw.name}`);
+
+      const games = await draw.getGames({
+        transaction,
+      });
+
+      for (const game of games) {
+        existing.games.push({
+          id: game.id,
+          visualCode: game.visualCode,
+        });
+        await game.destroy({ transaction });
+      }
+
       await draw.destroy({ transaction });
-      existed = true;
+      draw = undefined;
+      existing.existed = true;
     }
 
     if (!drawCode) {
@@ -106,36 +138,42 @@ export class DrawTournamentProcessor {
       throw new Error('Sub draw not found');
     }
 
-    draw = new DrawTournament({
-      id: drawId,
-      subeventId: subEvent.id,
-      visualCode: visualDraw.Code,
-      name: visualDraw.Name,
-      size: visualDraw.Size,
-      type:
-        visualDraw.TypeID === XmlDrawTypeID.Elimination
-          ? DrawType.KO
-          : visualDraw.TypeID === XmlDrawTypeID.RoundRobin ||
-              visualDraw.TypeID === XmlDrawTypeID.FullRoundRobin
-            ? DrawType.POULE
-            : DrawType.QUALIFICATION,
-    });
+    if (!draw) {
+      draw = new DrawTournament();
+    }
+    if (drawId) {
+      draw.id = drawId;
+    }
+
+    draw.subeventId = subEvent.id;
+    draw.visualCode = visualDraw.Code;
+    draw.name = visualDraw.Name;
+    draw.size = visualDraw.Size;
+    draw.type =
+      visualDraw.TypeID === XmlDrawTypeID.Elimination
+        ? DrawType.KO
+        : visualDraw.TypeID === XmlDrawTypeID.RoundRobin ||
+            visualDraw.TypeID === XmlDrawTypeID.FullRoundRobin
+          ? DrawType.POULE
+          : DrawType.QUALIFICATION;
 
     await draw.save({ transaction });
 
     let gameJobIds = [];
     // if we request to update the draws or the event is new we need to process the matches
-    if (job.data.updateMatches || !existed) {
+    if (options.updateMatches || !existing.existed) {
       gameJobIds = await this.processGames(
         job.data.eventCode,
         draw.visualCode,
         draw,
         job.data.rankingSystemId,
         job.data.transactionId,
+        existing.games,
+        options,
       );
     }
 
-    if (job.data.updateStanding || !existed) {
+    if (options.updateStanding || !existing.existed) {
       // also schedule a standing job
       const standingJob = await this._syncQueue.add(Sync.ProcessSyncTournamentDrawStanding, {
         transactionId: job.data.transactionId,
@@ -153,9 +191,13 @@ export class DrawTournamentProcessor {
     draw: DrawTournament,
     rankingSystemId: string,
     transactionId: string,
+    games: { id: string; visualCode: string }[],
+    options: {
+      deleteMatches?: boolean;
+    },
   ) {
     const transaction = await this._transactionManager.getTransaction(transactionId);
-    const matches = await this._visualService.getMatches(eventCode, drawCode, true);
+    const matches = await this._visualService.getGames(eventCode, drawCode, true);
 
     // remove all sub events in this event that are not in the visual to remove stray data
     const dbGames = await draw.getGames({
@@ -174,13 +216,15 @@ export class DrawTournamentProcessor {
     // queue the new sub events
     for (const match of matches) {
       // update sub events
-      const matchJob = await this._syncQueue.add(Sync.ProcessSyncTournamentMatch, {
+      const matchJob = await this._syncQueue.add(Sync.ProcessSyncTournamentGame, {
         transactionId,
         subEventId: draw.subeventId,
         eventCode,
         rankingSystemId,
         drawId: draw.id,
         gameCode: match.Code,
+        gameId: games?.find((r) => r.visualCode === `${match.Code}`)?.id,
+        options,
       });
 
       await this._transactionManager.addJob(transactionId, matchJob);
